fix: move sleep to have all code in src
